Split post-execution test
